backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.0
drf-yasg>=1.21
python-dotenv>=1.0
gunicorn>=20.1
djangorestframework-simplejwt>=5.3
Pillow>=10.0

